// src/app/helpers/enums/measure-unit.ts
export enum MeasureUnit {
  CUP = "CUP",
  DOZEN = "DOZEN",
  DROP = "DROP",
  GRAM = "GRAM",
  HUNDRED = "HUNDRED",
  KILOGRAM = "KILOGRAM",
  KIT = "KIT",
  LITER = "LITER",
  MILLIGRAM = "MILLIGRAM",
  MILLILITER = "MILLILITER",
  PACK = "PACK",
  PINCH = "PINCH",
  SERVING = "SERVING",
  SLICE = "SLICE",
  TEASPOON = "TEASPOON",
  TABLESPOON = "TABLESPOON",
  TABLET = "TABLET",
  UNIT = "UNIT",
}